fix clear output

+++ alexp/src/main/java/alexp/macrobase/Benchmark.java
@@ -80,7 +80,7 @@ public class Benchmark {
         if (options.has(outputOption)) {
             outputDir = outputOption.value(options);
             includeInliers = options.has(includeInliersOutputOption);
-            if (options.has(clearOutputOption)) {
+            if (options.has(clearOutputOption) && Files.exists(Paths.get(outputDir))) {
                 try {
                     FileUtils.cleanDirectory(new File(outputDir));
                 } catch (IOException ex) {
